requirements.txt
annotated-types==0.7.0
async-timeout==4.0.3
autocommand==2.2.2
backports.tarfile==1.2.0
bcrypt==4.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
colorama==0.4.6
commlib-py==0.11.4
crashtest==0.4.1
cryptography==43.0.3
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.9
dulwich==0.22.7
fabric==3.2.2
fastjsonschema==2.21.1
filelock==3.16.1
hiredis==3.1.0
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
installer==0.7.0
invoke==2.2.0
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==3.12.1
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
packaging==24.1
paho-mqtt==1.6.1
paramiko==3.5.0
pika==1.3.2
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.1
poetry-core==2.0.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-memcached==1.62
PyYAML==6.0.1
RapidFuzz==3.11.0
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
SecretStorage==3.3.3
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
tomli==2.0.1
tomlkit==0.13.2
trove-classifiers==2025.1.15.22
typeguard==4.3.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
virtualenv==20.29.1
wheel==0.45.1
wrapt==1.16.0
zipp==3.19.2
